updating for consistency with conversion to doubles for ssc_number_t

+++ shared/lib_util.cpp
@@ -1048,9 +1048,9 @@ bool util::translate_schedule(int tod[8760], const matrix_t<float> &wkday, const
 			for (size_t h = 0; h<24; h++)
 			{
 				if (is_weekday)
-					tod[i] = (size_t)wkday.at(m, h);
+					tod[i] = (int)wkday.at(m, h);
 				else
-					tod[i] = (size_t)wkend.at(m, h);
+					tod[i] = (int)wkend.at(m, h);
 
 				if (tod[i] < min_val) tod[i] = min_val;
 				if (tod[i] > max_val) tod[i] = max_val;
